% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MST_Main.R
\docType{data}
\name{SimX}
\alias{SimX}
\title{Simulated Data}
\format{An object of class \code{list} of length 100.}
\usage{
data(SimX)
}
\description{
This is a simulated dataset with 100 observations from 4 by 3 matrix skew-t distribution.
}
\keyword{data}
